backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
web3==6.11.0
openai==1.3.5
pytest==7.4.3
httpx==0.25.1
sqlalchemy==2.0.23
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
pillow==10.1.0
aiohttp==3.8.6